+++ src/game/Object/Player.cpp
@@ -5152,223 +5152,6 @@ void Player::setFactionForRace(uint8 race)
     setFaction(getFactionForRace(race));
 }
 
-/**
- * @brief Gets the player's current reputation rank with a faction.
- *
- * @param faction The faction identifier to query.
- * @return The current reputation rank.
- */
-ReputationRank Player::GetReputationRank(uint32 faction) const
-{
-    FactionEntry const* factionEntry = sFactionStore.LookupEntry(faction);
-    return GetReputationMgr().GetRank(factionEntry);
-}
-
-// Calculate total reputation percent player gain with quest/creature level
-int32 Player::CalculateReputationGain(ReputationSource source, int32 rep, int32 faction, uint32 creatureOrQuestLevel, bool noAuraBonus)
-{
-    float percent = 100.0f;
-
-    float repMod = noAuraBonus ? 0.0f : (float)GetTotalAuraModifier(SPELL_AURA_MOD_REPUTATION_GAIN);
-
-    // faction specific auras only seem to apply to kills
-    if (source == REPUTATION_SOURCE_KILL)
-    {
-        repMod += GetTotalAuraModifierByMiscValue(SPELL_AURA_MOD_FACTION_REPUTATION_GAIN, faction);
-    }
-
-    percent += rep > 0 ? repMod : -repMod;
-
-    float rate;
-    switch (source)
-    {
-        case REPUTATION_SOURCE_KILL:
-            rate = sWorld.getConfig(CONFIG_FLOAT_RATE_REPUTATION_LOWLEVEL_KILL);
-            break;
-        case REPUTATION_SOURCE_QUEST:
-            rate = sWorld.getConfig(CONFIG_FLOAT_RATE_REPUTATION_LOWLEVEL_QUEST);
-            break;
-        case REPUTATION_SOURCE_SPELL:
-        default:
-            rate = 1.0f;
-            break;
-    }
-
-    if (rate != 1.0f && creatureOrQuestLevel <= MaNGOS::XP::GetGrayLevel(getLevel()))
-    {
-        percent *= rate;
-    }
-
-    if (percent <= 0.0f)
-    {
-        return 0;
-    }
-
-    // Multiply result with the faction specific rate
-    if (const RepRewardRate* repData = sObjectMgr.GetRepRewardRate(faction))
-    {
-        float repRate = 0.0f;
-        switch (source)
-        {
-            case REPUTATION_SOURCE_KILL:
-                repRate = repData->creature_rate;
-                break;
-            case REPUTATION_SOURCE_QUEST:
-                repRate = repData->quest_rate;
-                break;
-            case REPUTATION_SOURCE_SPELL:
-                repRate = repData->spell_rate;
-                break;
-        }
-
-        // for custom, a rate of 0.0 will totally disable reputation gain for this faction/type
-        if (repRate <= 0.0f)
-        {
-            return 0;
-        }
-
-        percent *= repRate;
-    }
-
-    return int32(sWorld.getConfig(CONFIG_FLOAT_RATE_REPUTATION_GAIN) * rep * percent / 100.0f);
-}
-
-// Calculates how many reputation points player gains in victim's enemy factions
-void Player::RewardReputation(Unit* pVictim, float rate)
-{
-    if (!pVictim || pVictim->GetTypeId() == TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // used current difficulty creature entry instead normal version (GetEntry())
-    ReputationOnKillEntry const* Rep = sObjectMgr.GetReputationOnKillEntry(((Creature*)pVictim)->GetCreatureInfo()->Entry);
-
-    if (!Rep)
-    {
-        return;
-    }
-
-    uint32 repFaction1 = Rep->repfaction1;
-    uint32 repFaction2 = Rep->repfaction2;
-
-    // Championning tabard reputation system
-    // Aura 57818 is a hidden aura common to tabards allowing championning.
-    if (pVictim->GetMap()->IsNonRaidDungeon() && HasAura(57818))
-    {
-        MapEntry const* storedMap = sMapStore.LookupEntry(GetMapId());
-        InstanceTemplate const* instance = ObjectMgr::GetInstanceTemplate(GetMapId());
-        Item const* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, EQUIPMENT_SLOT_TABARD);
-        if (storedMap && instance && pItem)
-        {
-            ItemPrototype const* pProto = pItem->GetProto();// Checked on load
-            // The required MinLevel for the tabard to work is related to the item level of the tabard
-            if ((instance->levelMin + 1 >= pProto->ItemLevel || !GetMap()->IsRegularDifficulty())
-                    // For ItemLevel == 75 (or 85) need to check expansion
-                    && (pProto->ItemLevel == 75 && storedMap->Expansion() == EXPANSION_WOTLK))
-            {
-                if (uint32 tabardFactionID = pItem->GetProto()->RequiredReputationFaction)
-                {
-                    repFaction1 = tabardFactionID;
-                    repFaction2 = tabardFactionID;
-                }
-            }
-        }
-    }
-
-    if (repFaction1 && (!Rep->team_dependent || GetTeam() == ALLIANCE))
-    {
-        int32 donerep1 = CalculateReputationGain(REPUTATION_SOURCE_KILL, Rep->repvalue1, repFaction1, pVictim->getLevel());
-        donerep1 = int32(donerep1 * rate);
-        FactionEntry const* factionEntry1 = sFactionStore.LookupEntry(repFaction1);
-        uint32 current_reputation_rank1 = GetReputationMgr().GetRank(factionEntry1);
-        if (factionEntry1 && current_reputation_rank1 <= Rep->reputation_max_cap1)
-        {
-            GetReputationMgr().ModifyReputation(factionEntry1, donerep1);
-        }
-
-        // Wiki: Team factions value divided by 2
-        if (factionEntry1 && Rep->is_teamaward1)
-        {
-            FactionEntry const* team1_factionEntry = sFactionStore.LookupEntry(factionEntry1->team);
-            if (team1_factionEntry)
-            {
-                GetReputationMgr().ModifyReputation(team1_factionEntry, donerep1 / 2);
-            }
-        }
-    }
-
-    if (repFaction2 && (!Rep->team_dependent || GetTeam() == HORDE))
-    {
-        int32 donerep2 = CalculateReputationGain(REPUTATION_SOURCE_KILL, Rep->repvalue2, repFaction2, pVictim->getLevel());
-        donerep2 = int32(donerep2 * rate);
-        FactionEntry const* factionEntry2 = sFactionStore.LookupEntry(repFaction2);
-        uint32 current_reputation_rank2 = GetReputationMgr().GetRank(factionEntry2);
-        if (factionEntry2 && current_reputation_rank2 <= Rep->reputation_max_cap2)
-        {
-            GetReputationMgr().ModifyReputation(factionEntry2, donerep2);
-        }
-
-        // Wiki: Team factions value divided by 2
-        if (factionEntry2 && Rep->is_teamaward2)
-        {
-            FactionEntry const* team2_factionEntry = sFactionStore.LookupEntry(factionEntry2->team);
-            if (team2_factionEntry)
-            {
-                GetReputationMgr().ModifyReputation(team2_factionEntry, donerep2 / 2);
-            }
-        }
-    }
-}
-
-// Calculate how many reputation points player gain with the quest
-void Player::RewardReputation(Quest const* pQuest)
-{
-    // quest reputation reward/loss
-    for (int i = 0; i < QUEST_REPUTATIONS_COUNT; ++i)
-    {
-        if (!pQuest->RewRepFaction[i])
-        {
-            continue;
-        }
-
-        // No diplomacy mod are applied to the final value (flat). Note the formula (finalValue = DBvalue/100)
-        if (pQuest->RewRepValue[i])
-        {
-            int32 rep = CalculateReputationGain(REPUTATION_SOURCE_QUEST, pQuest->RewRepValue[i] / 100, pQuest->RewRepFaction[i], GetQuestLevelForPlayer(pQuest), true);
-
-            if (FactionEntry const* factionEntry = sFactionStore.LookupEntry(pQuest->RewRepFaction[i]))
-            {
-                GetReputationMgr().ModifyReputation(factionEntry, rep);
-            }
-        }
-        else
-        {
-            uint32 row = ((pQuest->RewRepValueId[i] < 0) ? 1 : 0) + 1;
-            uint32 field = abs(pQuest->RewRepValueId[i]);
-
-            if (const QuestFactionRewardEntry* pRow = sQuestFactionRewardStore.LookupEntry(row))
-            {
-                int32 repPoints = pRow->rewardValue[field];
-
-                if (!repPoints)
-                {
-                    continue;
-                }
-
-                repPoints = CalculateReputationGain(REPUTATION_SOURCE_QUEST, repPoints, pQuest->RewRepFaction[i], GetQuestLevelForPlayer(pQuest));
-
-                if (const FactionEntry* factionEntry = sFactionStore.LookupEntry(pQuest->RewRepFaction[i]))
-                {
-                    GetReputationMgr().ModifyReputation(factionEntry, repPoints);
-                }
-            }
-        }
-    }
-
-    // TODO: implement reputation spillover
-}
-
 // Player::UpdateHonorKills moved to HonorMgr::UpdateKills (2026-05-12); thin delegating wrapper lives inline in Player.h.
 
 /// Calculate the amount of honor gained based on the victim
